+++ bgplot/features/operations/use_cases/distances.py
@@ -4,6 +4,6 @@ from ....entities import Point
 
 
 def distance_between_two_points(point_1: Point, point_2: Point) -> float:
-    distance: float = np.linalg.norm(point_1-point_2)
+    distance: float = np.linalg.norm(point_1-point_2).astype(float)
 
     return distance